[SC-105] refactor: 채팅방 메세지 인풋 컴포넌트 분리
