+++ src/api/chat/getChatRoomData.tsx
@@ -1,20 +1,20 @@
 import axios, {AxiosResponse} from 'axios';
-import { PostReissue } from './postReissue';
+import { PostReissue } from '../user/postReissue';
 
 const BASE_URL = 'https://www.codin.co.kr/api';
 
-export const GetVoteData = async (accessToken:string, page:number): Promise<any> => {
+export const GetChatRoomData = async (accessToken:string): Promise<any> => {
     axios.defaults.withCredentials = true;
     try{
         const response: AxiosResponse<any> = await axios.get(
-            `${BASE_URL}/posts/category?postCategory=POLL&page=${page}`,
+            `${BASE_URL}/chatroom`,
            { headers: {
              Authorization: ` ${accessToken}`
           }}
         );
         console.log('토큰:', accessToken);
         console.log(response.data);
-        return response.data;
+        return response;
       } catch (error: any) {
         if (error.response) {
           const { status, data } = error.response;
@@ -23,7 +23,7 @@ export const GetVoteData = async (accessToken:string, page:number): Promise<any>
             console.error('401 Unauthorized: 토큰이 유효하지 않거나 만료되었습니다.');
             const refreshToken = localStorage.getItem('refresh-token');
             localStorage.setItem('accessToken', refreshToken);
-            GetVoteData(refreshToken, page);
+            GetChatRoomData(refreshToken);
 
           }
         } else if (error.request) {
