+++ src/api/comment/getComments.tsx
@@ -1,20 +1,16 @@
 import axios, {AxiosResponse} from 'axios';
-import { PostReissue } from './postReissue';
+import { PostReissue } from '../user/postReissue';
 
 const BASE_URL = 'https://www.codin.co.kr/api';
 
-export const PostComments = async (accessToken:string, postId:string | string[], content:string, anonymous:boolean): Promise<any> => {
+export const GetComments = async (accessToken:string, postId:string | string[]): Promise<any> => {
     axios.defaults.withCredentials = true;
     try{
-        const response: AxiosResponse<any> = await axios.post(
-            `${BASE_URL}/comments/${postId}`,
-            {
-                content : content,
-                anonymous: anonymous
-            },
+        const response: AxiosResponse<any> = await axios.get(
+            `${BASE_URL}/comments/post/${postId}`,
            { headers: {
              Authorization: ` ${accessToken}`
-          }},
+          }}
         );
         console.log('토큰:', accessToken);
         console.log(response.data);
@@ -27,7 +23,7 @@ export const PostComments = async (accessToken:string, postId:string | string[],
             console.error('401 Unauthorized: 토큰이 유효하지 않거나 만료되었습니다.');
             const refreshToken = localStorage.getItem('refresh-token');
             localStorage.setItem('accessToken', refreshToken);
-            PostComments(refreshToken, postId, content, anonymous);
+            GetComments(refreshToken, postId);
 
           }
         } else if (error.request) {
