+++ src/api/like/postLike.tsx
@@ -1,16 +1,20 @@
 import axios, {AxiosResponse} from 'axios';
-import { PostReissue } from './postReissue';
+import { PostReissue } from '../user/postReissue';
 
 const BASE_URL = 'https://www.codin.co.kr/api';
 
-export const GetComments = async (accessToken:string, postId:string | string[]): Promise<any> => {
+export const PostLike = async (accessToken:string, likeType:string, id:string): Promise<any> => {
     axios.defaults.withCredentials = true;
     try{
-        const response: AxiosResponse<any> = await axios.get(
-            `${BASE_URL}/comments/post/${postId}`,
+        const response: AxiosResponse<any> = await axios.post(
+            `${BASE_URL}/likes`,
+            {
+                likeType : likeType,
+                id: id
+            },
            { headers: {
              Authorization: ` ${accessToken}`
-          }}
+          }},
         );
         console.log('토큰:', accessToken);
         console.log(response.data);
@@ -23,7 +27,7 @@ export const GetComments = async (accessToken:string, postId:string | string[]):
             console.error('401 Unauthorized: 토큰이 유효하지 않거나 만료되었습니다.');
             const refreshToken = localStorage.getItem('refresh-token');
             localStorage.setItem('accessToken', refreshToken);
-            GetComments(refreshToken, postId);
+            PostLike(refreshToken, likeType, id);
 
           }
         } else if (error.request) {
